Fix merge conflict concerning int types and prinft variants

+++ includes/server.h
@@ -9,11 +9,11 @@ typedef struct s_server
 {
 	volatile sig_atomic_t				*stop;
 	t_client							*head;
-	int32_t								max_fd;		
-	uint8_t								next_id;
+	int									max_fd;		
+	uint64_t								next_id;
 	fd_set								active;
 	fd_set								readfds;
-	int32_t								socket;
+	int									socket;
 	uint16_t							errno_code;
 	struct sockaddr_in					address;
 	uint32_t							ip;
@@ -28,13 +28,13 @@ t_error		create_socket(t_server *server);
 t_error		bind_address(t_server *server);
 t_error		set_passive_socket(t_server *server);
 void		accept_new_client(t_server *server);
-void		handle_client(t_server *server, const int32_t fd);
-void		remove_client(t_server *server, const int32_t fd);
-void		handle_message(t_server *server, const int32_t fd, const char *r);
-void		broadcast(t_server *server, const int16_t except);
+void		handle_client(t_server *server, const int fd);
+void		remove_client(t_server *server, const int fd);
+void		handle_message(t_server *server, const int fd, const char *r);
+void		broadcast(t_server *server, const int except);
 void		add_to_list(t_server *server, t_client *client);
 void		remove_from_list(t_server *server, t_client *client);
 void		release_ressources(t_server *server);
-t_client	*find_client(t_server *server, const int32_t fd);
+t_client	*find_client(t_server *server, const int fd);
 
 #endif
